requirements.txt
PyPDF2
huggingface_hub
langchain
langchain-community
transformers
neo4j
faiss-cpu
google-generativeai
python-dotenv
filelock
fsspec
packaging
pyyaml
requests
tqdm
typing-extensions
python-dotenv
langchain-core
langchain-text-splitters
langsmith
pydantic
SQLAlchemy
tenacity
jsonpatch
httpx
orjson
requests-toolbelt
zstandard
annotated-types
pydantic-core
charset-normalizer
idna
urllib3
certifi
greenlet
anyio
httpcore
h11
jsonpointer
sniffio
aiohttp
dataclasses-json
pydantic-settings
httpx-sse
llmsherpa
pandas
sentence-transformers
tensorflow
keras
tensorflow-text
tensorflow-io